Split `config.py` into small modules
